Huge push! This merge includes: Cleanup Basic (one-box-to-another) collision A WIP dynamic AABBTree class for broadphase collision with insertion, removal, and balancing. It's WIP because querying the tree for actually collision stuff to happen is Not Yet Implemented.
Enemy class updated for testing.
Some spring cleaning here and there. :)

Basic collision check function works.

Now that collision works, I'm getting rid of RectangularBoundaryCollison.
Though I've modified EnemiesOld, I'm still keeping it around so I can easily reference it instead of having to have it open in another editor or window or something.

Revert "Now that collision works, I'm getting rid of RectangularBoundaryCollison."

This reverts commit 5806be0a3638e6fe7f365ec91281c460132799b2.

Now that basic collision works, I decided to get rid of RectangularBoundaryCollision and do some random spring cleaning!
I had no idea there were spaghetti-like #include lines everywhere. D'oh.

Added logic to enclose two AABBs into one larger AABB.
Now I can try to make a tree structure and traverse it for a broadphase.

quick changes

Insertion into the AABBTree works.

useless comment removed.

Fix awkward spacing.

useless comment

The Dynamic AABBTree is working!!! (So far).

Added balance() method to tree.
More bugs need fixing... should make a free() method.

AllocateNode method!
Tree balancing and addition is FIXED!!!
Small change to removal may need investigating, though.

Fixed a silly bug that messed up shooting.

Added 1000 enemies and their own collision boxes to the tree, running without drawing the AABBs. Let's see how this works on a laptop...

+++ cplusplusGameExperimental/cplusplusGameExperimental/State.cpp
@@ -25,7 +25,7 @@ void State::add(ObjectInfo & o)
 
 void State::add(vector<ObjectInfo> & v)
 {
-   for (int i = 0; i < v.size(); i++) {
+   for (int i = 0; i < (int) v.size(); i++) {
       v[i].drawIndex = current.size() * 4;
       current.push_back(&v[i]);
       previous.push_back(v[i]);
